Added a number of classes.

+++ JavaUtils/src/id/ac/ub/filkom/rendicahya/utils/matrix/interfaces/MatrixTwoValueMapper.java
@@ -2,11 +2,11 @@ package id.ac.ub.filkom.rendicahya.utils.matrix.interfaces;
 
 public interface MatrixTwoValueMapper {
 
-    byte[][] operate(byte[][] a, int b, int c);
+    byte[][] operate(byte[][] a, byte b, byte c);
 
     short[][] operate(short[][] a, short b, short c);
 
-    int[][] operate(int[][] a, short b, short c);
+    int[][] operate(int[][] a, int b, int c);
 
     long[][] operate(long[][] a, long b, long c);
